requirements.txt
Flask==0.11.1
gunicorn==19.6.0
Pillow==3.4.2